Version 0.1.26 - Leveling/XP system, 8ball rooms, adding UI, fixing bugs
- Leveling system has been implemented - You can now receive levels and XP
for playing 8 ball & chess games. You earn more if you win, and there’s
no increase in XP wagerless/wagered for chess.

- Soon enough, there will be new ways to earn XP, & at a specific
level(s)/milestones you will be rewarded with badges, cosmetics, and
perhaps even airdrops. Currently we are developing a profile system, and
a & notification system so you can know when you have leveled up. The
profile system will be available for anyone to find other players, check
their stats, and will be available/interoperable with our future apps.
More info on this on our development board.

- Showcase 8 ball games being played on game lobbies.  - We wanted to show
users who are playing the 8 ball challenge, even if it’s single player.
We created a couple endpoints, to create this mechanism. When you
initiate a game, it will show on the front page for 10 minutesI am
working towards implementing multiplayer as soon as possible to get our
next game, Drawades, available as soon as possible.

- Changed game lobby numbering/indexing to game controller icon due to
problems with side effects with the two endpoints fetching the  rooms

- optimized fetching authentication data on component mounting (a bit)

- Fixed a few bugs related to chess moves and fetching chess data

- fixed a bug that gets the ball stuck in rare exceptions in 8 ball

- fixed or changed a few un-notable UI improvements